修复: AVI 帧块 size 字段误写 8 字节导致视频无法打开
- 根因: SKData.Size 是 long，_writer.Write(data.Size) 调用 Write(long)
  把 AVI chunk size 字段写成 8 字节（规范为 4 字节），每帧数据错位 4 字节，
  短视频被解码器容错，完整视频全部损坏（ffmpeg 报 mjpeg overread）
- 修复: 强制写 (int)data.Size
- 新增回归测试: 验证每个帧块为 4 字节 size + 紧跟 JPEG SOI
- 验证: 完整 1080p@30fps 7087 帧视频 ffmpeg 零错误解码（317MB, 3:56）

+++ UstViz.Rendering/Video/MjpegAviVideoWriter.cs
@@ -139,7 +139,7 @@ public sealed class MjpegAviVideoWriter : IVideoWriter
         uint chunkOffset = (uint)(_stream.Position - _moviDataStart);
 
         WriteFourCc("00dc");
-        _writer.Write(data.Size);
+        _writer.Write((int)data.Size); // 注意：SKData.Size 是 long，必须转 int（AVI size 字段 4 字节）
         _writer.Write(data.ToArray());
 
         // 2 字节对齐
@@ -211,6 +211,7 @@ public sealed class MjpegAviVideoWriter : IVideoWriter
 
     public void Dispose() => Finish();
 }
+
